Set UDP packet length to 1252 in all situations
This is to address an inconsistency between the behavior with mahi-mahi
vs tperf:

- tperf has a flag `canIgnorePathMTU` that defaults to true, and makes
it so that `conn_.udpSendPacketLen` is updated from 1232 to 1252 after
receiving connection info from the peer.

- In mvfst-rl this flag is currently false, and as a result
`conn_.udpSendPacketLen` stays at 1232. Setting this flag to true does
not solve this issue because the increase to 1252 also is gated by a
check on the peer's packet size, that must be <= 1252. But with
mahi-mahi packets are actually of size 1500 bytes (the MTU size used by
mahi-mahi) and as a result this check fails, and
`conn_.udpSendPacketLen` remains at 1232.

Additional notes:

- This problem is new, and was caused by the recent update of mvfst.
With the older mvfst the packet size used was also 1252 in mvfst-rl, and
thus there was no discrepancy vs tperf.

- This discrepancy probably does not matter much in practice, since the
congestion window is measured in bytes and thus the congestion control
algorithm should in theory work with arbitrary packet sizes. But having
a consistent packet size in mahi-mahi vs tperf helps with debugging,
which is the main motivation for this commit.

- Setting all packet sizes to 1500 in mvfst (in order to match the MTU
size used by mahi-mahi) causes connection errors.